feat: reduce package size to from ~180mb to ~15mb (WIP)
By not bundling the `aws-cdk-lib` dependency into the Wing SDK we
can reduce the package size dramatically.

When compiling to the `awscdk` target, the compiler will check if
`aws-cdk-lib` is installed (with the right version). If not, a nice
message will be printed with instructions.

> In the future we can automatically install aws-cdk-lib just in time, but I think this is a reasonable experience for the time being.

TODO:

- [ ] Fix SDK ESLint rules to not yell at us
- [ ] Add CLI logic to check for `aws-cdk-lib` installation when using the `awscdk` target
