+++ libs/wingsdk/package.json
@@ -50,6 +50,7 @@
     "@vitest/coverage-c8": "^0.29.2",
     "@winglang/jsii-docgen": "file:../../apps/jsii-docgen",
     "@winglang/wing-api-checker": "file:../../apps/wing-api-checker",
+    "aws-cdk-lib": "^2.77.0",
     "aws-sdk-client-mock": "^2.0.1",
     "aws-sdk-client-mock-jest": "^2.0.1",
     "cdktf-cli": "0.15.2",
@@ -92,7 +93,6 @@
     "@azure/core-paging": "^1.4.0",
     "@azure/identity": "3.1.3",
     "@azure/storage-blob": "12.14.0",
-    "aws-cdk-lib": "^2.64.0",
     "cdktf": "0.15.2",
     "constructs": "~10.1.228",
     "cron-parser": "^4.8.1",
@@ -121,7 +121,6 @@
     "@azure/core-paging",
     "@azure/identity",
     "@azure/storage-blob",
-    "aws-cdk-lib",
     "cdktf",
     "cron-parser",
     "debug",
