+++ libs/wingsdk/src/target-tf-gcp/app.ts
@@ -3,7 +3,8 @@ import { Bucket } from "./bucket";
 import { GoogleProvider } from "../.gen/providers/google/provider";
 import { RandomProvider } from "../.gen/providers/random/provider";
 import { BUCKET_FQN } from "../cloud";
-import { AppProps as CdktfAppProps, CdktfApp } from "../core";
+import { AppProps as CdktfAppProps } from "../core";
+import { CdktfApp } from "../shared-tf/app";
 
 /**
  * GCP App props.
